Remove SDL_gtk
